move ast and compiler under `syntax`
