--- src/scenic/ast/__init__.py
@@ -1,1 +1,0 @@
-from .ast import AST, EgoAssign, New
