--- src/scenic/compiler/__init__.py
@@ -1,1 +1,0 @@
-from .compiler import compileScenicAST
