--- src/scenic/parser/__init__.py
@@ -1,1 +1,0 @@
-from .parser import parse_file, parse_string
