+++ src/scenic/syntax/compiler.py
@@ -1,6 +1,6 @@
 import ast
 from typing import Tuple, List
-import scenic.ast as s
+import scenic.syntax.ast as s
 
 # exposed functions
