+++ tests/parser/helper.py
@@ -1,5 +1,5 @@
 from typing import Any
-from scenic.parser import parse_string
+from scenic.syntax.parser import parse_string
 
 def parse_string_helper(source: str) -> Any:
     "Parse string and return Scenic AST"
