+++ tests/parser/test_new.py
@@ -1,7 +1,7 @@
 import ast
 
 from .helper import parse_string_helper
-import scenic.ast as s
+import scenic.syntax.ast as s
 
 class TestNew:
     def test_basic(self):
